Using the XMLHttpRequest() object:
This example creates a new XMLHttpRequest object and sets it up to make a GET request to the URL https://example.com/data.json. The response type is set to JSON, and an onload function is defined to log the response to the console. An onerror function is defined to handle any errors that occur. Finally, the request is sent using the send() method.